Fix anchor build issues

+++ programs/clmm/src/constants.rs
@@ -1,7 +1,7 @@
 use anchor_lang::prelude::*;
 
 #[constant]
-pub const ANCHOR_SIZE: usize = 8;
+pub const ANCHOR_SIZE: u32 = 8;
 
 #[constant]
 pub const SEED: &str = "anchor";
@@ -25,10 +25,10 @@ pub const TICK_ARRAY_BITMAP_SEED: &str = "tick_array_bitmap";
 pub const FEE_RATE_DENOMINATOR_VALUE: u32 = 1_000_000;
 
 #[constant]
-pub const TICK_ARRAY_SIZE: usize = 60;
+pub const TICK_ARRAY_SIZE: u32 = 60;
 
 #[constant]
-pub const TICK_ARRAY_BITMAP_SIZE: usize = 128;
+pub const TICK_ARRAY_BITMAP_SIZE: u32 = 128;
 
 #[constant]
 pub const TICK_MAX: i32 = 443636;
@@ -37,7 +37,7 @@ pub const TICK_MIN: i32 = -443636;
 
 /// The minimum value that can be returned from #get_sqrt_price_at_tick. Equivalent to get_sqrt_price_at_tick(TICK_MIN)
 #[constant]
-pub const SQRT_PRICE_X64_MIN: u128 = 4295048016;
+pub const SQRT_PRICE_X64_MIN: u128 = 4295048016u128;
 /// The maximum value that can be returned from #get_sqrt_price_at_tick. Equivalent to get_sqrt_price_at_tick(TICK_MAX)
 #[constant]
-pub const SQRT_PRICE_X64_MAX: u128 = 79226673521066979257578248091;
+pub const SQRT_PRICE_X64_MAX: u128 = 79226673521066979257578248091u128;
